feat: Upgrade to Electron 12
At the recommendation of @dbluhm, I have taken a look at migrating the
work on the ARIES Toolbox from using Electron 9 to Electron 12.
@burdettadam found and recommended the use of
mubaidr/vue-electron-template, which has been paramount in this upgrade.

This change involved moving the minimal amount of code from
hyperledger/aries-toolbox into the vue electron template until things
started to work. Once the UI was able to show up, and a connection to
an instance of hyperledger/aries-cloudagent-python could be established
and feature list discovered, success was determined in this endeavor.

There do appear to be a couple of console errors at this time that may
need cleaning up before this change can be considered complete. Once
everything is working, this should fix hyperledger/aries-toolbox#237.

Successful connection via loca.lt has been determined. It is assumed
that loca.lt uses Let's Encrypt certificates as described in the
aforementioned issue.

I plan to spend Monday working on fixes for the remaining console errors
that I saw earlier, clean up any unnecessary code changes, and get a PR
ready for the upstream repo.

Signed-off-by: Colton Wolkins (Indicio work address) <colton@indicio.tech>