fix: Warn instead of error if Agent has no public DID
An error that I noticed getting thrown was that `msg.result` was
undefined and did not have the property `did` associated with it. Better
than an error, I have made it so the code only runs if msg.result is
defined. `share.fetch_dids();` is also in this block because it would
not run anyways if the line above it threw an exception.

If `msg.result` is `undefined`, we now return a warning that no public
did was found.

Signed-off-by: Colton Wolkins (Indicio work address) <colton@indicio.tech>